[UPD] add function in system to finish creation of ennemy

+++ server/game_engine/include/Rectangle.hpp
@@ -13,13 +13,13 @@
 namespace game_engine {
     struct Rectangle
     {
-        Rectangle(size_t _x, size_t _y, size_t _L, size_t _l)
+        Rectangle(int _x, int _y, int _L, int _l)
         : x(_x), y(_y), L(_L), l(_l)
         {}
-        size_t x;
-        size_t y;
-        size_t L;
-        size_t l;
+        int x;
+        int y;
+        int L;
+        int l;
     };
 }
